Make FInAT interface not depend on FIAT interface
